requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asyncio==3.4.3
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.104.1
gunicorn==21.2.0
h11==0.14.0
idna==3.4
packaging==23.2
prometheus-client==0.19.0
pydantic==2.5.1
pydantic_core==2.14.3
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
tapo==0.1.3
typing_extensions==4.8.0
uvicorn==0.24.0.post1
